Cleaned up `AvroConversionHelper`
