+++ hudi-spark-datasource/hudi-spark/src/test/scala/org/apache/hudi/TestAvroConversionHelper.scala
@@ -18,12 +18,12 @@
 
 package org.apache.hudi
 
-import java.time.LocalDate
-
 import org.apache.avro.Schema
 import org.apache.avro.generic.GenericData
 import org.apache.spark.sql.catalyst.expressions.GenericRow
 import org.scalatest.{FunSuite, Matchers}
+
+import java.time.LocalDate
 
 class TestAvroConversionHelper extends FunSuite with Matchers {
 
@@ -42,7 +42,7 @@ class TestAvroConversionHelper extends FunSuite with Matchers {
 
   test("Logical type: date") {
     val schema = new Schema.Parser().parse(dateSchema)
-    val convertor = AvroConversionHelper.createConverterToRow(schema, AvroConversionUtils.convertAvroSchemaToStructType(schema))
+    val convertor = AvroConversionUtils.createConverterToRow(schema, AvroConversionUtils.convertAvroSchemaToStructType(schema))
 
     val dateOutputData = dateInputData.map(x => {
       val record = new GenericData.Record(schema) {{ put("date", x) }}
